+++ proactive-p2p-structured/proactive-p2p-structured-can/src/main/java/org/objectweb/proactive/extensions/p2p/structured/validator/can/DefaultAnycastConstraintsValidator.java
@@ -45,7 +45,7 @@ public final class DefaultAnycastConstraintsValidator extends
     }
 
     public final boolean validatesKeyConstraints(Zone zone) {
-        for (int i = 0; i < super.key.size(); i++) {
+        for (byte i = 0; i < super.key.size(); i++) {
             // if coordinate is null we skip the test
             if (super.key.getElement(i) != null) {
                 // the specified overlay does not contains the key
